requirements.txt
streamlit
langchain
openai
langchain_community
python-dotenv
tiktoken
chromadb
ipykernel
pypdf
biopython
